Move facility specific loading code to `FacilityPickerView`
