+++ app/src/main/java/org/simple/clinic/facilitypicker/FacilityPickerEffect.kt
@@ -1,3 +1,18 @@
 package org.simple.clinic.facilitypicker
 
+import org.simple.clinic.facility.Facility
+import java.time.Duration
+
 sealed class FacilityPickerEffect
+
+data class FetchCurrentLocation(
+    val updateInterval: Duration,
+    val timeout: Duration,
+    val discardOlderThan: Duration
+) : FacilityPickerEffect()
+
+data class LoadFacilitiesWithQuery(val query: String) : FacilityPickerEffect()
+
+object LoadTotalFacilityCount : FacilityPickerEffect()
+
+data class ForwardSelectedFacility(val facility: Facility) : FacilityPickerEffect()
