+++ app/src/main/java/org/simple/clinic/facilitypicker/FacilityPickerUi.kt
@@ -1,3 +1,11 @@
 package org.simple.clinic.facilitypicker
 
-interface FacilityPickerUi
+import org.simple.clinic.facility.change.FacilityListItem
+
+interface FacilityPickerUi {
+  fun showProgressIndicator()
+  fun hideProgressIndicator()
+  fun showToolbarWithSearchField()
+  fun showToolbarWithoutSearchField()
+  fun updateFacilities(facilityItems: List<FacilityListItem>)
+}
