--- app/src/main/java/org/simple/clinic/registration/facility/RegistrationFacilitySelectionUi.kt
@@ -1,11 +1,0 @@
-package org.simple.clinic.registration.facility
-
-import org.simple.clinic.facility.change.FacilityListItem
-
-interface RegistrationFacilitySelectionUi {
-  fun showProgressIndicator()
-  fun hideProgressIndicator()
-  fun showToolbarWithSearchField()
-  fun showToolbarWithoutSearchField()
-  fun updateFacilities(facilityItems: List<FacilityListItem>)
-}
