+++ app/src/main/java/org/simple/clinic/registration/facility/RegistrationFacilitySelectionUpdate.kt
@@ -1,7 +1,6 @@
 package org.simple.clinic.registration.facility
 
 import com.spotify.mobius.Next
-import com.spotify.mobius.Next.next
 import com.spotify.mobius.Update
 import org.simple.clinic.mobius.dispatch
 
@@ -12,10 +11,6 @@ class RegistrationFacilitySelectionUpdate : Update<RegistrationFacilitySelection
       event: RegistrationFacilitySelectionEvent
   ): Next<RegistrationFacilitySelectionModel, RegistrationFacilitySelectionEffect> {
     return when (event) {
-      is LocationFetched -> next(model.locationFetched(event.update))
-      is FacilitiesFetched -> next(model.queryChanged(event.query).facilitiesLoaded(event.facilities))
-      is RegistrationFacilitySearchQueryChanged -> dispatch(LoadFacilitiesWithQuery(event.query))
-      is TotalFacilityCountLoaded -> next(model.facilityCountLoaded(event.count))
       is RegistrationFacilityClicked -> dispatch(OpenConfirmFacilitySheet(event.facility))
       is RegistrationFacilityConfirmed -> {
         val updatedEntry = model.ongoingEntry.withFacilityUuid(event.facilityUuid)
